Split out the Room DAOs
