+++ app/src/main/java/me/banes/chris/tivi/calls/TrendingCall.kt
@@ -25,6 +25,7 @@ import io.reactivex.Maybe
 import io.reactivex.Single
 import me.banes.chris.tivi.data.TiviShow
 import me.banes.chris.tivi.data.TiviShowDao
+import me.banes.chris.tivi.data.TrendingDao
 import me.banes.chris.tivi.data.TrendingEntry
 import me.banes.chris.tivi.util.AppRxSchedulers
 import me.banes.chris.tivi.util.DatabaseTxRunner
@@ -33,13 +34,14 @@ import javax.inject.Inject
 class TrendingCall @Inject constructor(
         databaseTxRunner: DatabaseTxRunner,
         showDao: TiviShowDao,
+        private val trendingDao: TrendingDao,
         tmdb: Tmdb,
         trakt: TraktV2,
         schedulers: AppRxSchedulers)
     : PaginatedTraktCall<TrendingShow>(databaseTxRunner, showDao, tmdb, trakt, schedulers) {
 
     override fun createData(page: Int?): Flowable<List<TiviShow>> {
-        return if (page == null) showDao.trendingShows() else showDao.trendingShowsPage(page)
+        return if (page == null) trendingDao.trendingShows() else trendingDao.trendingShowsPage(page)
     }
 
     override fun networkCall(page: Int): Single<List<TrendingShow>> {
@@ -57,21 +59,21 @@ class TrendingCall @Inject constructor(
     }
 
     override fun lastPageLoaded(): Single<Int> {
-        return showDao.getLastTrendingPage()
+        return trendingDao.getLastTrendingPage()
     }
 
     override fun deleteEntries() {
-        showDao.deleteTrendingShows()
+        trendingDao.deleteTrendingShows()
     }
 
     override fun deletePage(page: Int) {
-        showDao.deleteTrendingShowsPageSync(page)
+        trendingDao.deleteTrendingShowsPageSync(page)
     }
 
     override fun saveEntry(show: TiviShow, page: Int, order: Int) {
         assert(show.id != null)
         val entry = TrendingEntry(showId = show.id!!, page = page, pageOrder = order)
-        showDao.insertTrending(entry)
+        trendingDao.insertTrending(entry)
     }
 
 }
