+++ app/src/main/java/me/banes/chris/tivi/data/TiviShowDao.kt
@@ -21,9 +21,7 @@ import android.arch.persistence.room.Insert
 import android.arch.persistence.room.OnConflictStrategy
 import android.arch.persistence.room.Query
 import android.arch.persistence.room.Update
-import io.reactivex.Flowable
 import io.reactivex.Maybe
-import io.reactivex.Single
 
 @Dao
 interface TiviShowDao {
@@ -44,59 +42,5 @@ interface TiviShowDao {
 
     @Update(onConflict = OnConflictStrategy.REPLACE)
     fun updateShow(shows: TiviShow)
-
-    /**
-     * Trending shows
-     */
-
-    @Query("SELECT * FROM shows " +
-            "INNER JOIN trending_shows ON trending_shows.show_id = shows.id " +
-            "ORDER BY page, page_order")
-    fun trendingShows(): Flowable<List<TiviShow>>
-
-    @Query("SELECT * FROM shows " +
-            "INNER JOIN trending_shows ON trending_shows.show_id = shows.id " +
-            "WHERE page = :page " +
-            "ORDER BY page_order")
-    fun trendingShowsPage(page: Int): Flowable<List<TiviShow>>
-
-    @Insert(onConflict = OnConflictStrategy.IGNORE)
-    fun insertTrending(show: TrendingEntry): Long
-
-    @Query("DELETE FROM trending_shows WHERE page = :page")
-    fun deleteTrendingShowsPageSync(page: Int = 0)
-
-    @Query("DELETE FROM trending_shows")
-    fun deleteTrendingShows()
-
-    @Query("SELECT MAX(page) from trending_shows")
-    fun getLastTrendingPage(): Single<Int>
-
-    /**
-     * Popular shows
-     */
-
-    @Query("SELECT * FROM shows " +
-            "INNER JOIN popular_shows ON popular_shows.show_id = shows.id " +
-            "ORDER BY page, page_order")
-    fun popularShows(): Flowable<List<TiviShow>>
-
-    @Query("SELECT * FROM shows " +
-            "INNER JOIN popular_shows ON popular_shows.show_id = shows.id " +
-            "WHERE page = :page " +
-            "ORDER BY page_order")
-    fun popularShowsPage(page: Int): Flowable<List<TiviShow>>
-
-    @Insert(onConflict = OnConflictStrategy.IGNORE)
-    fun insertPopularShows(show: PopularEntry): Long
-
-    @Query("DELETE FROM popular_shows WHERE page = :page")
-    fun deletePopularShowsPageSync(page: Int = 0)
-
-    @Query("DELETE FROM popular_shows")
-    fun deletePopularShows()
-
-    @Query("SELECT MAX(page) from popular_shows")
-    fun getLastPopularPage(): Single<Int>
 
 }
